chore(common): add bundling support
